requirements.txt
diskcache
llama-cpp-python
